[luci] Separate DeadNodeQueryService (#2356)
This will separate DeadNodeQueryServiceImpl class to own file

ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
